refactor: merge state provider
